feat:#245 커스텀 훅 분리
