+++ src/app/(client)/report/constants/chart.constant.ts
@@ -36,4 +36,4 @@ export const MACRONUTRIENT_GOAL_MAP = {
   FAT: 'dailyFatGoal'
 } as const;
 
-export const INITIAL_NUTRITION_VALUE = { calories: 0, carbohydrate: 0, protein: 0, fat: 0 };
+export const INITIAL_NUTRITION_VALUE = { calories: -1, carbohydrate: 0, protein: 0, fat: 0 };
